Added button in projects page
Added button in place of <a href...></a> in projects for "Return to main page".
Fixed the font size difference from index and projects.
Made the button to return to main page in projects page and the footer in same line.
Essentially removed footer tag.
Removed the two <hr> tags before footer.
Added margin of 7px at left to <a> tags in links part.